+++ deepinv/loss/adversarial/__init__.py
@@ -3,6 +3,9 @@ from .consistency import (
     SupAdversarialDiscriminatorLoss,
     UnsupAdversarialGeneratorLoss,
     UnsupAdversarialDiscriminatorLoss,
+    MultiOperatorUnsupAdversarialGeneratorLoss,
+    MultiOperatorUnsupAdversarialDiscriminatorLoss,
 )
-from .uair import UAIRGeneratorLoss
+from .uair import UAIRGeneratorLoss, UAIRDiscriminatorLoss
 from .base import DiscriminatorLoss, GeneratorLoss, DiscriminatorMetric
+from .mo import MultiOperatorMixin