+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ New Features
 ^^^^^^^^^^^^
 
 - Added :class:`deepinv.physics.SaltPepperNoise` for impulse noise (:gh:`472` by `Thomas Moreau`_).
+- Add multi-operator unsup adversarial loss (:gh:`409` by `Andrew Wang`_)
+
 
 Changed
 ^^^^^^^
@@ -23,6 +25,7 @@ Fixed
 - Fix latex rendering in `deepinv.utils.plotting.config_matplotlib`  (:gh:`452` by `Romain Vo`_)
 - Get rid of unnecessary file system writes in `get_freer_gpu` (:gh:`468` by `Jérémy Scanvic`_)
 
+- Fix UAIR loss (:gh:`409` by `Andrew Wang`_)
 
 v0.3
 ----------------